+++ src/main/resources/lib/ssb/dataset/klass/municipalities.ts
@@ -11,6 +11,13 @@ import {
   fromMunicipalityWithCodeCache,
   fromMunicipalityWithNameCache,
 } from '/lib/ssb/cache/cache'
+import {
+  MunicipalCode,
+  MunicipalityChange,
+  MunicipalityChangeList,
+  MunicipalityWithCounty,
+  RequestWithCode,
+} from '/lib/types/municipalities'
 import { type DataSource } from '/site/mixins/dataSource'
 
 /**
@@ -213,41 +220,4 @@ function getMunicipalityChanges(): MunicipalityChangeList {
 
 export function removeCountyFromMunicipalityName(municipalityName: string): string {
   return municipalityName.split('(')[0].trim()
-}
-
-export interface MunicipalityChangeList {
-  codeChanges: Array<MunicipalityChange>
-}
-
-export interface MunicipalityChange {
-  oldCode: string
-  oldName: string
-  oldShortName?: string
-  newCode: string
-  newName: string
-  newShortName?: string
-  changeOccurred: string
-}
-
-export interface RequestWithCode extends XP.Request {
-  code: string
-}
-
-export interface MunicipalCode {
-  code: string
-  parentCode: string
-  level: string
-  name: string
-  shortName: string
-  presentationName: string
-}
-
-export interface MunicipalityWithCounty {
-  code: string
-  displayName: string
-  county: {
-    name: string
-  }
-  path: string
-  changes?: Array<MunicipalityChange>
 }
