+++ src/main/resources/lib/ssb/parts/highcharts/highchartsData.ts
@@ -1,7 +1,8 @@
 import { Content } from '/lib/xp/content'
 import { PreliminaryData, TbmlDataUniform } from '/lib/types/xmlParser'
 import { JSONstat } from '/lib/types/jsonstat-toolkit'
-import { RowValue, getRowValue } from '/lib/ssb/utils/utils'
+import { RowValue } from '/lib/types/util'
+import { getRowValue } from '/lib/ssb/utils/utils'
 import { seriesAndCategoriesFromHtmlTable } from '/lib/ssb/parts/highcharts/data/htmlTable'
 import { seriesAndCategoriesFromJsonStat } from '/lib/ssb/parts/highcharts/data/statBank'
 import { seriesAndCategoriesFromTbml } from '/lib/ssb/parts/highcharts/data/tbProcessor'
