+++ src/main/resources/lib/ssb/parts/publicationArchive.ts
@@ -2,7 +2,7 @@ import { type Content, get, query, QueryDsl, ContentsResult } from '/lib/xp/cont
 import { pageUrl } from '/lib/xp/portal'
 import { type StatisticInListing } from '/lib/ssb/dashboard/statreg/types'
 import { getAllStatisticsFromRepo } from '/lib/ssb/statreg/statistics'
-import { calculatePeriodRelease, type Release, getPreviousReleases } from '/lib/ssb/utils/variantUtils'
+import { calculatePeriodRelease, getPreviousReleases } from '/lib/ssb/utils/variantUtils'
 import {
   type SubjectItem,
   getMainSubjects,
@@ -19,6 +19,8 @@ import {
 import { fromPartCache } from '/lib/ssb/cache/partCache'
 import { isEnabled } from '/lib/featureToggle'
 import * as util from '/lib/util'
+import { PublicationItem, PublicationResult } from '/lib/types/partTypes/publicationArchive'
+import { Release } from '/lib/types/variants'
 import { type Article, type OmStatistikken, type Statistics } from '/site/content-types'
 
 export function getPublications(
@@ -292,24 +294,4 @@ function getSecondaryMainSubject(
     return acc
   }, [])
   return secondaryMainSubjects
-}
-
-export interface PublicationResult {
-  total: number
-  publications: Array<PublicationItem>
-}
-
-export interface PublicationItem {
-  title: string
-  period?: string
-  preface: string
-  url: string
-  publishDate: string
-  publishDateHuman: string | undefined
-  contentType: string
-  articleType: string
-  mainSubjectId: string
-  mainSubject: string
-  secondaryMainSubjects: Array<string>
-  appName: string
 }
