+++ src/main/resources/lib/ssb/parts/table.ts
@@ -28,6 +28,14 @@ import * as util from '/lib/util'
 import { getDataset } from '/lib/ssb/dataset/dataset'
 import { datasetOrUndefined } from '/lib/ssb/cache/cache'
 import { getRowValue } from '/lib/ssb/utils/utils'
+import {
+  DatasourceHtmlTable,
+  HtmlTable,
+  HtmlTableRaw,
+  HtmlTableRowRaw,
+  BodyCell,
+  TableView,
+} from '/lib/types/partTypes/table'
 import { type DataSource } from '/site/mixins/dataSource'
 import { type Table } from '/site/content-types'
 
@@ -306,51 +314,4 @@ function getNoterefsHeader(row: TableCellUniform): Array<string> {
     return acc
   }, [])
   return noteRefs
-}
-
-export interface TableView {
-  caption?: Title
-  thead: Array<TableRowUniform>
-  tbody: Array<TableRowUniform>
-  tfoot: {
-    footnotes: Array<Note>
-    correctionNotice: string
-  }
-  tableClass: string
-  noteRefs: Array<string>
-  sourceList: Source | TableSourceList | undefined
-}
-
-export type TableSourceList = Array<Source>
-
-interface DatasourceHtmlTable {
-  html: string | undefined
-  footnoteText: Array<string>
-}
-
-interface HtmlTableRaw {
-  table: {
-    tbody: {
-      tr: Array<HtmlTableRowRaw>
-    }
-  }
-}
-interface HtmlTableRowRaw {
-  td: Array<number | string | PreliminaryData>
-}
-
-export interface HtmlTable {
-  table: {
-    thead: {
-      tr: {
-        th: Array<number | string>
-      }
-    }
-    tbody: {
-      tr: Array<BodyCell>
-    }
-  }
-}
-interface BodyCell {
-  td: Array<number | string>
 }
