+++ src/main/resources/lib/types/partTypes/publicationArchive.ts
@@ -1,5 +1,3 @@
-import { PublicationResult } from '/lib/ssb/parts/publicationArchive'
-
 export interface PublicationArchiveProps {
   title: string
   ingress: string
@@ -21,4 +19,24 @@ export interface PublicationArchiveProps {
 export interface Dropdown {
   id: string
   title: string
+}
+
+export interface PublicationResult {
+  total: number
+  publications: Array<PublicationItem>
+}
+
+export interface PublicationItem {
+  title: string
+  period?: string
+  preface: string
+  url: string
+  publishDate: string
+  publishDateHuman: string | undefined
+  contentType: string
+  articleType: string
+  mainSubjectId: string
+  mainSubject: string
+  secondaryMainSubjects: Array<string>
+  appName: string
 }
