+++ src/main/resources/lib/types/util.ts
@@ -7,3 +7,5 @@ interface UtilDataLibrary {
 }
 
 export declare const data: UtilDataLibrary
+
+export type RowValue = number | string
