+++ src/main/resources/site/parts/highmap/highmap.ts
@@ -1,7 +1,8 @@
 import { get as getContentByKey, getAttachmentStream, type ByteSource, type Content } from '/lib/xp/content'
 import { getComponent, getContent } from '/lib/xp/portal'
 import { readText } from '/lib/xp/io'
-import { isNumber, type RowValue } from '/lib/ssb/utils/utils'
+import { isNumber } from '/lib/ssb/utils/utils'
+import { type RowValue } from '/lib/types/util'
 import { render } from '/lib/enonic/react4xp'
 import { type PreliminaryData, type XmlParser, type TbmlDataUniform, TableRowUniform } from '/lib/types/xmlParser'
 import * as util from '/lib/util'
